feat: better hooks
