+++ src/lib/index.ts
@@ -1,2 +1,1 @@
-// Reexport your entry components here
-export * from "./tinybase/hooks.svelte";
+export * from "./tinybase/hooks/index.js";
